Modifications for updating and deletion

+++ src/product/product.service.ts
@@ -10,8 +10,8 @@ import { Product } from './entity/product.entity';
 import { CreateProductDto } from './dto/createProduct.dto';
 import { Subcategory } from '../subcategory/entity/subcategory.entity';
 import { User } from '../user/entity/user.entity';
-import { CurrentUserService } from '../common/currentUser.service';
 import { ProductDto } from './dto/product.dto';
+import { CurrentUserDto } from '../common/currentUser.dto';
 
 @Injectable()
 export class ProductService {
@@ -22,13 +22,15 @@ export class ProductService {
     private readonly subcategoryRepository: Repository<Subcategory>,
     @InjectRepository(User)
     private readonly userRepository: Repository<User>,
-    private readonly currentUserService: CurrentUserService,
   ) {}
 
-  async create(createProductDto: CreateProductDto, user): Promise<Product> {
+  async create(
+    createProductDto: CreateProductDto,
+    user: CurrentUserDto
+  ): Promise<Product> {
     const { name, description, price, mainImageUrl, subcategoryId } =
       createProductDto;
-      const sellerId: string = user.userId;
+    const sellerId: string = user.userId;
     try {
       const existingProduct = await this.productRepository.findOne({
         where: { name },
@@ -79,18 +81,40 @@ export class ProductService {
     }
   }
 
-  async remove(id: string) {
+  async remove(id: string, user: CurrentUserDto) {
     const product = await this.productRepository.findOne({
       where: { id },
+      relations: ['seller'],
     });
+
+    if (product.seller.id !== user.userId && user.role.name !== 'admin') {
+      throw new ConflictException(
+        'You are not authorized to delete this product',
+      );
+    }
 
     if (!product) {
       throw new NotFoundException(`No product found with this id:${id}`);
     }
 
-    this.productRepository.remove(product);
+    await this.productRepository.remove(product);
   }
-  async update(updateProduct: CreateProductDto, id: string): Promise<Product> {
+  async update(
+    updateProduct: CreateProductDto,
+    id: string,
+    user: CurrentUserDto,
+  ): Promise<Product> {
+    const product = await this.productRepository.findOne({
+      where: { id },
+      relations: ['seller'],
+    });
+
+    if (product.seller.id !== user.userId && user.role.name !== 'admin') {
+      throw new ConflictException(
+        'You are not authorized to update this product',
+      );
+    }
+
     const newProduct = await this.productRepository.preload({
       id: id,
       ...updateProduct,
@@ -120,8 +144,31 @@ export class ProductService {
           category: product.subcategory.category.name,
         };
       });
+    } catch (error) {
+      console.log(error);
+      throw new InternalServerErrorException(error);
+    }
+  }
 
-    } catch(error) {
+  async getUserProducts(user: CurrentUserDto): Promise<ProductDto[]> {
+    try {
+      const products = await this.productRepository.find({
+        where: { seller: { id: user.userId } },
+        relations: ['subcategory', 'subcategory.category', 'seller'],
+      });
+
+      return products.map((product): ProductDto => {
+        return {
+          id: product.id,
+          name: product.name,
+          description: product.description,
+          price: product.price,
+          mainImageUrl: product.mainImageUrl,
+          subcategory: product.subcategory.name,
+          category: product.subcategory.category.name,
+        };
+      });
+    } catch (error) {
       console.log(error);
       throw new InternalServerErrorException(error);
     }
